+++ xarray/core/alignment.py
@@ -105,7 +105,8 @@ def align(*objects, **kwargs):
                 try:
                     # GH1434
                     # dtype is lost after obj.indexes[dim]
-                    # problem originates in  Indexes.__getitem__ in coordinates.py
+                    # problem originates in  Indexes.__getitem__
+                    # in coordinates.py
                     index = obj.indexes[dim]
                 except KeyError:
                     unlabeled_dim_sizes[dim].add(obj.sizes[dim])